exoplayer code out of Adapter.
